:art: setup presentational and container component
